+++ src/lvmcryo/handlers/valve.py
@@ -10,6 +10,7 @@ from __future__ import annotations
 
 import asyncio
 import logging
+import warnings
 from dataclasses import dataclass, field
 from datetime import UTC, datetime
 
@@ -20,6 +21,7 @@ from rich.progress import TaskID
 from lvmopstools.clu import CluClient
 from lvmopstools.retrier import Retrier
 
+from lvmcryo import RetrierWarning
 from lvmcryo.config import get_internal_config
 from lvmcryo.handlers.thermistor import ThermistorHandler
 from lvmcryo.tools import cancel_task, get_fake_logger, ln2_estops
@@ -40,7 +42,13 @@ __all__ = [
 ]
 
 
-@Retrier(max_attempts=3, delay=1, timeout=10)
+def log_retrier_error(error: Exception):
+    """Logs an error from a retrier."""
+
+    warnings.warn(f"Error in retrier: {error}. Retrying ...", RetrierWarning)
+
+
+@Retrier(max_attempts=3, delay=1, timeout=10, on_retry=log_retrier_error)
 async def outlet_info(actor: str, outlet: str) -> dict[str, Any]:
     """Retrieves outlet information from the NPS."""
 
@@ -52,7 +60,7 @@ async def outlet_info(actor: str, outlet: str) -> dict[str, Any]:
     return cmd.replies.get("outlet_info")
 
 
-@Retrier(max_attempts=3, delay=1, timeout=30)
+@Retrier(max_attempts=3, delay=1, timeout=30, on_retry=log_retrier_error)
 async def valve_on_off(
     actor: str,
     outlet_name: str,
@@ -125,7 +133,7 @@ async def valve_on_off(
     return
 
 
-@Retrier(max_attempts=3, delay=1, timeout=10)
+@Retrier(max_attempts=3, delay=1, timeout=10, on_retry=log_retrier_error)
 async def cancel_nps_threads(actor: str, thread_id: int | None = None):
     """Cancels a script thread in an NPS.
 
@@ -144,7 +152,7 @@ async def cancel_nps_threads(actor: str, thread_id: int | None = None):
         await client.send_command(actor, command_string)
 
 
-@Retrier(max_attempts=3, delay=1, timeout=60)
+@Retrier(max_attempts=3, delay=1, timeout=60, on_retry=log_retrier_error)
 async def close_all_valves(config: Configuration | None = None, dry_run: bool = False):
     """Closes all the outlets."""